Test unitaires pour getMethods et base introspectable pour faciliter les tests
